1_prepare_data/docker/requirements.txt
jsonlines
pillow